pyproject.toml
[tool.poetry]
name = "it_requests"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.20.0"
fastapi = "^0.88.0"
dependency-injector = "^4.41.0"
aiopg = "^1.4.0"
pydantic = "^1.10.4"
python-dotenv = "^0.21.0"
yoyo-migrations = "^8.2.0"
httpx = "^0.23.2"
aiofiles = "^23.1.0"
python-multipart = "^0.0.6"
aiogram = "3.0.0b7"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.0.0"
isort = "^5.11.4"
black = "^22.12.0"
docformatter = "^1.5.1"
add-trailing-comma = "^2.4.0"
flake8 = "^6.0.0"
mypy = "^0.991"
safety = "^2.3.5"
bandit = "^1.7.4"
itsdangerous = "^2.1.2"
markupsafe = "^2.1.1"
sphinx = "^4.4.0"
sphinx-pydantic = "^0.1.1"
sphinx-rtd-theme = "^1.0.0"
pytest-asyncio = "^0.20.3"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
